[SPMD] Split the SPMD flow to tracing, profiling and optimization stages
[ghstack-poisoned]
