+++ spmd/compiler/bucketing_strategies.py
@@ -1,0 +1,7 @@
+from enum import auto, Enum
+
+
+class BucketingStrategy(Enum):
+    FIXED = auto()
+    VARIABLE = auto()
+    CONSTANT = auto()
